clean up plexcron
